Изменен модуль для отрисовки миниатюр
