+++ js/main.js
@@ -1,3 +1,7 @@
 
-import {picturesElement} from './thumbnails.js';
-console.log (picturesElement);
+import {createThumbnails} from './create-thumbnails.js';
+import {createPhotos} from './create-photos.js';
+
+const pictures = createPhotos();
+createThumbnails(pictures);
+
